requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.*
google.generativeai==0.3.*
google-auth==2.*
google-auth-httplib2==0.*
google-api-python-client==2.*
google-auth-oauthlib==1.*
google-cloud-storage==2.*
beautifulsoup4==4.*
lxml==4.*